Исправление комментариев
